pyproject.toml
[project]
name = "agentic-python-coder"
version = "2.0.1"
description = "A lightweight Python coding agent that writes, executes, and iterates on code through natural language instructions"
readme = "README.md"
requires-python = ">=3.13,<3.14"
authors = [
    {name = "Stefan Szeider"}
]
license = {text = "Apache-2.0"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
keywords = ["ai", "coding", "agent", "llm", "code-generation", "automation"]
dependencies = [
    "langgraph>=1.0.4",
    "langchain-core>=1.1.0",
    "langchain-openai>=1.1.0",
    "langchain-anthropic>=1.2.0",
    "langchain-experimental>=0.4.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "jupyter-client>=8.6.3",
    "ipykernel>=6.30.1",
    "PyYAML>=6.0.3",
]

[project.scripts]
coder = "agentic_python_coder.cli:main"

[project.optional-dependencies]
test = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-watch>=4.2.0",
]
dev = [
    "ruff>=0.14.7",
    "mypy>=1.19.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = [
    "CPMPY/",
    "ZEBRA/",
    "PAPER-ASP/",
    "LOCAL/",
    "TRASH/",
    "CPBENCH/",
    ".mcp.json",
    "*.jsonl",
    "*_code.py",
    "examples/cpmpy/cpmpy_*.md",
    "examples/clingo/clingo_*.md",
    "RELEASE_NOTES_*.md",
]

[tool.hatch.build.targets.wheel]
packages = ["coder/src/agentic_python_coder"]

[tool.hatch.build.targets.wheel.force-include]
"coder/prompts" = "agentic_python_coder/prompts"

[tool.pytest.ini_options]
pythonpath = ["coder/src"]

[tool.ruff]
target-version = "py313"
line-length = 88
exclude = ["CPBENCH", "LOCAL", "TRASH", "examples"]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true


